HF3D: Add support for erasing config without 'full erase'
When a Makefile flag FLASH_CONFIG_ERASE=yes is defined,
the generated .hex/.bin file will overwrite the config
area in the flash, effectively erasing the config.

This flag is supposed to be used with the snapshot images.

During development the config flash structure is often
changed, which invalidates any existing content in the flash.
This will show up as corrupted config in the CLI.

The best way to avoid surprise corruption is to erase the
config with the development images, and force the user
to reconfigure or reload the config from a text dump.

Release images should _NEVER_ use this flag!