HF3D: getMotorCount() returns 1 or 2, never 0
In many places in the code, it is assumed that there is
at least one motor. Bad things will happen if getMotorCount()
returns zero, which is the case when the mixer is not configured.

This commit makes getMotorCount() always return at least one.
In practice it will return 1 or 2, as MAX_SUPPORTED_MOTORS is 2.